refactor(Rubros form): move component to its own directory
